docs: fix accordion route spelling
