+++ docs/app/components/blocks/accordion/page.module.css
